Menu Bug Fix
Fixed a bizarre design flaw in how the navigation menus were being rendered.

The menu was defined in base_aashe.html, but when this was edited, the
changes were not being reflected. It appeared the menu div was being rendered
twice. This was circumvented by commenting out the inclusion of sitemap.html,
which was strange since that does not touch the top menu at all.

What was happening was the html loaded, then in main.js a function was
called to populate the drop-downs. It did so by searching for the class
'menu-block-3', which was ONLY present in sitemap.html.

That class identifier was moot, it was only being used for this function
in the javascript and not referenced in the CSS, so the fix was to simply
remove it from sitemap.html and put it instead into the menu div of
base_aashe.html.

Now, to edit the menu, one simply edits the lines in that script as one
was attempting to do previously.