Fix Templatetag Queries
Wrap objects.get() calls within templatetags in try/except to fix broken
tests on travis for SCPD and Bulletin. Was not creating the Menu objects
inside tests, so loading those tags with get(pk=1) was returning
errors becuase that object was never created, causing the whole page request
to fail.